backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "providers*", "routers*"]

[project]
name = "securerelay-backend"
version = "0.1.0"
description = "SecureRelay backend — privacy layer for LLM-assisted pentesting"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "gliner>=0.2.8",
    "onnxruntime>=1.18.0",
    "pyyaml>=6.0.1",
    "huggingface-hub>=0.23.0",
    "sqlcipher3-binary>=0.5.0",
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "keyring>=25.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]
